requirements.txt
instagrapi
yt-dlp
google-api-python-client
python-dotenv
